Move convert_to_boolean into app.__init__
This is the only place that it's used and it's the only thing in the
helpers module.
