+++ app/__init__.py
@@ -2,12 +2,12 @@ import re
 import os
 
 from flask import Flask
+from flask._compat import string_types
 from flask.ext.bootstrap import Bootstrap
 from flask.ext.sqlalchemy import SQLAlchemy
 from werkzeug.contrib.fixers import ProxyFix
 
 from config import config
-from .helpers import convert_to_boolean
 
 
 bootstrap = Bootstrap()
@@ -43,3 +43,27 @@ def config_attrs(config):
     """Returns config attributes from a Config object"""
     p = re.compile('^[A-Z_]+$')
     return filter(lambda attr: bool(p.match(attr)), dir(config))
+
+
+def convert_to_boolean(value):
+    """Turn strings to bools if they look like them
+
+    Truthy things should be True
+    >>> for truthy in ['true', 'on', 'yes', '1']:
+    ...   assert convert_to_boolean(truthy) == True
+
+    Falsey things should be False
+    >>> for falsey in ['false', 'off', 'no', '0']:
+    ...   assert convert_to_boolean(falsey) == False
+
+    Other things should be unchanged
+    >>> for value in ['falsey', 'other', True, 0]:
+    ...   assert convert_to_boolean(value) == value
+    """
+    if isinstance(value, string_types):
+        if value.lower() in ['t', 'true', 'on', 'yes', '1']:
+            return True
+        elif value.lower() in ['f', 'false', 'off', 'no', '0']:
+            return False
+
+    return value
