--- app/helpers.py
@@ -1,25 +1,0 @@
-from flask._compat import string_types
-
-
-def convert_to_boolean(value):
-    """Turn strings to bools if they look like them
-
-    Truthy things should be True
-    >>> for truthy in ['true', 'on', 'yes', '1']:
-    ...   assert convert_to_boolean(truthy) == True
-
-    Falsey things should be False
-    >>> for falsey in ['false', 'off', 'no', '0']:
-    ...   assert convert_to_boolean(falsey) == False
-
-    Other things should be unchanged
-    >>> for value in ['falsey', 'other', True, 0]:
-    ...   assert convert_to_boolean(value) == value
-    """
-    if isinstance(value, string_types):
-        if value.lower() in ['t', 'true', 'on', 'yes', '1']:
-            return True
-        elif value.lower() in ['f', 'false', 'off', 'no', '0']:
-            return False
-
-    return value
